Allow residence verification by zipcode
Notes:

Once the new code has been deployed to production and DB migrations
have been run, a rake task needs to be executed in order to import the
zipcodes from the CSV file to the DB

`bin/rake db:zipcodes RAILS_ENV=production`

After that, any user will be verified by just introducing a valid
zipcode.

For the error message when introducing a zipcode that is not in the
list, the default error message shown will be this one:
https://github.com/codefornl/consul/blob/v0.19.2/config/locales/nl/verification.yml#L63
Feel free to adjust it as your convenience for your locale.

You can do it by adding the file:
`config/locales/custom/nl/verification.yml`
with the content

```
nl:
  verification:
    residence:
      new:
        error_not_allowed_postal_code: "CUSTOM ERROR MESSAGE"
```